backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.76.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.28.1
python-multipart==0.0.22
pymupdf==1.26.7
python-docx==1.2.0
